Change MDT mixins to functions, refactor packages
