+++ nbmolviz/methods/__init__.py
@@ -1,0 +1,6 @@
+from .atoms import *
+from .atomgroups import *
+from .molecules import *
+from .method import *
+from .residues import *
+from .trajectory import *
