--- nbmolviz/mixins/__init__.py
@@ -1,4 +1,0 @@
-from .mixins import *
-from .atoms import *
-from .molecule import *
-from .method import *
