--- nbmolviz/viewer/__init__.py
@@ -1,8 +1,0 @@
-def toplevel(o):
-    __all__.append(o.__name__)
-    return o
-__all__ = []
-
-from .common import *
-from .viewer2d import *
-from .viewer3d import *
