+++ nbmolviz/viewers/viewer3d.py
@@ -22,12 +22,11 @@ from moldesign import utils
 from moldesign.helpers import VolumetricGrid, colormap
 
 from ..base.drivers3d import MolViz_3DMol
-from . import toplevel, ColorMixin
+from . import ColorMixin
 
 
 # Right now hard-coded to use 3DMol driver - will need to be configurable when there's more than one
 # In theory, the only tie to the specific driver should be the class that we inherit from
-@toplevel
 class GeometryViewer(MolViz_3DMol, ColorMixin):
     """
     Viewer for static and multiple-frame geometries
