+++ nbmolviz/widgets/__init__.py
@@ -7,7 +7,4 @@ from .geombuilder import *
 from .selection import *
 from .symmetry import *
 
-from . import trajectory
-from . import orbitals
-from . import configurator
